moved authorized_acoount to abstract package
